fixing mvc booking and fixing mvc user booking order

+++ application/controllers/KamarController.php
@@ -9,6 +9,7 @@ class KamarController extends CI_Controller
 		$this->load->model('m_user');
 		$this->load->model('m_register');
 		$this->load->model('m_kamar');
+		$this->load->model('m_booking');
 		$this->load->helper('form');
 		$this->load->helper('url');
 		$this->load->library('form_validation');
@@ -16,36 +17,84 @@ class KamarController extends CI_Controller
 	}
 	public function user()
 	{
-		$this->load->view('header/header_user');
-		$this->load->view('dashboard/dashboard_user');
-		$this->load->view('footer/footer');
+		$id=$this->session->id;
+		$data['user']=$this->m_user->userprofile($id);
+		if($this->session->userdata('level') == 'user')
+		// if($sess_data['level'] == 'user' ) 
+			
+		{
+			$this->load->view('header/header_user');
+			$this->load->view('dashboard/dashboard_user');
+			$this->load->view('footer/ft_footer');
+			
+		} else {
+			redirect('login');
+		}
 	}
 
+	public function user_inbox()
+	{
+		$id=$this->session->id;
+		$data['user']=$this->m_booking->tampil_data_join_inbox($id);
+		// echo json_encode($data);die();
+		if($this->session->userdata('level') == 'user')
+		// if($sess_data['level'] == 'user' ) 
+			
+		{
+			$this->load->view('header/header_user');
+			$this->load->view('user/inbox',$data);
+			$this->load->view('footer/ft_footer');
+			
+		} else {
+			redirect('login');
+		}
+	}
+	
 	public function owner()
 	{
-		$this->load->view('header/header_owner');
-		$this->load->view('dashboard/dashboard_user');
-		$this->load->view('footer/ft_footer');
+		if($this->session->userdata('level') == 'owner')
+		{
+			$this->load->view('header/header_owner');
+			$this->load->view('dashboard/dashboard_user');
+			$this->load->view('footer/ft_footer');
+		}else{
+			redirect('login');
+		}
 	}
 
 	public function admin()
 	{
-		$this->load->view('header/header_admin');
-		$this->load->view('dashboard/dashboard_user');
-		$this->load->view('footer/ft_footer');
+		if($this->session->userdata('level') == 'admin')
+		{
+			$this->load->view('header/header_admin');
+			$this->load->view('dashboard/dashboard_user');
+			$this->load->view('footer/ft_footer');
+		}else{
+			redirect('login');
+		}
 	}
 	public function info_kamar()
 	{
-		$this->load->view("header/header_admin");
-		$data['kamar']=$this->m_kamar->tampil_data()->result();
-		$this->load->view("info_kamar",$data);
-		$this->load->view("footer/footer");
+		if($this->session->userdata('level') == 'admin')
+		{
+			$this->load->view("header/header_admin");
+			$data['kamar']=$this->m_kamar->tampil_data()->result();
+			$this->load->view("info_kamar",$data);
+			$this->load->view("footer/footer");
+		}else{
+			redirect('login');
+		}
 	}
 	public function tambah_kamar()
 	{
-		$this->load->view("header/header_admin");
-		$this->load->view('tambah_kamar');
-		$this->load->view("footer/footer");
+		if($this->session->userdata('level') == 'admin')
+		{
+			$this->load->view("header/header_admin");
+			$this->load->view('tambah_kamar');
+			$this->load->view("footer/footer");
+		}else{
+			redirect('login');
+		}
 	}
 	public function aksi_tambah()
 	{
@@ -64,17 +113,27 @@ class KamarController extends CI_Controller
 	}
 	public function edit($id)
 	{
-		$this->load->view("header/header_admin");
-		$where=array('id_kamar'=>$id);
-		$data['kamar']=$this->m_kamar->edit_data($where,'kamar')->result();
-		$this->load->view('edit_kamar',$data);
-		$this->load->view("footer/footer");
+		if($this->session->userdata('level') == 'admin')
+		{
+			$where=array('id_kamar'=>$id);
+			$data['kamar']=$this->m_kamar->edit_data($where,'kamar')->result();
+			$this->load->view("header/header_admin");
+			$this->load->view('edit_kamar',$data);
+			$this->load->view("footer/footer");
+		}else{
+			redirect('login');
+		}
 	}
 	public function hapus($id)
 	{
-		$where = array('id_kamar' => $id);
-		$this->m_kamar->hapus_data($where,'kamar');
-		redirect('info_kamar');
+		if($this->session->userdata('level') == 'admin')
+		{
+			$where = array('id_kamar' => $id);
+			$this->m_kamar->hapus_data($where,'kamar');
+			redirect('info_kamar');
+		}else{
+			redirect('login');
+		}
 	}
 	public function update()
 	{
